+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - (Feature) Delay Action
 - (Feature) MigrateMember Action
 - (Maintenance) Ensure Enum in the UpgradeMode Enum Schema
+- (Feature) Enforce UpgradeByReplace operation by default for ArangoDB from 3.12.2/3 to 3.12.4+
 
 ## [1.2.44](https://github.com/arangodb/kube-arangodb/tree/1.2.44) (2025-02-03)
 - (Maintenance) Kubernetes 1.31.1 libraries